BBlogFormat removed from namespace BlackBox
